add jenkinsfile and CNAME file
The current implementation of the site builder requires content
repo to be submited as a zip file for production using env vars
`CONTENT_SOURCE_TYPE=git` and `CONTENT_SOURCE_URL=repo-link.zip`
which is not considered a best practice. Due to this, a race
condition may occur during the build process if another commit
is made while the build is in progress which may lead a security
flow. To work around this,first, we checkout into `src` using
`checkoutToSubdirectory('src')`. Then, we use `local_folder`
mode using `CONTENT_SOURCE_TYPE=local_folder` and
`CONTENT_SOURCE_URL=..src` env vars, just like a local develop.
`local_folder` mode copies the website files from `src` into
`docs` because the site builder is incapable of copying it.
We checkout the builder into the `builder` directory and run
the `install` and `build` commands inside of that directory.
Finally, for publish stages, we copy the `out` directory into
the site repo, since the out directory is in the wrong repo.